# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customer/profile": {
      "get": {
        "operationId": "CustomerProfileController_getProfile[0]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current customer profile information",
        "tags": [
          "Customer - Profile"
        ]
      },
      "patch": {
        "operationId": "CustomerProfileController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update basic profile details",
        "tags": [
          "Customer - Profile"
        ]
      }
    },
...
    "/api/v1/customer/password": {
      "patch": {
        "operationId": "CustomerProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change account password",
        "tags": [
          "Customer - Profile"
        ]
      }
    },
    "/api/v1/customer/contact-info": {
      "get": {
        "operationId": "CustomerProfileController_getContactInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get secondary contact information",
        "tags": [
          "Customer - Profile"
        ]
      },
      "patch": {
        "operationId": "CustomerProfileController_updateContactInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update secondary contact information",
        "tags": [
          "Customer - Profile"
        ]
      }
    },
    "/api/v1/customer/preferences": {
      "get": {
        "operationId": "CustomerProfileController_getPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get notification and app preferences",
        "tags": [
          "Customer - Profile"
        ]
      },
      "patch": {
        "operationId": "CustomerProfileController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update notification and app preferences",
        "tags": [
          "Customer - Profile"
        ]
      }
    },
...
    "/api/v1/admin/expenses": {
      "post": {
        "operationId": "ExpensesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin - Expenses"
        ]
      },
      "get": {
        "operationId": "ExpensesController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin - Expenses"
        ]
      }
    },
    "/api/v1/admin/expenses/stats": {
      "get": {
        "operationId": "ExpensesController_getStats",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin - Expenses"
        ]
      }
    },
    "/api/v1/admin/expenses/{id}": {
      "get": {
        "operationId": "ExpensesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin - Expenses"
        ]
      },
      "patch": {
        "operationId": "ExpensesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin - Expenses"
        ]
      },
      "delete": {
        "operationId": "ExpensesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin - Expenses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Amman Communications API",
    "description": "API documentation and interactive endpoint specification for Amman Communications Business Platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter Admin JWT access token"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "contactNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          },
          "confirmNewPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword",
          "confirmNewPassword"
        ]
      },
      "UpdateContactInfoDto": {
        "type": "object",
        "properties": {
          "altContactName": {
            "type": "string"
          },
          "altPhoneNumber": {
            "type": "string"
          },
          "preferredContactMethod": {
            "type": "string"
          }
        }
      },
      "UpdatePreferencesDto": {
        "type": "object",
        "properties": {
          "emailNotifications": {
            "type": "boolean"
          },
          "smsAlerts": {
            "type": "boolean"
          },
          "whatsappUpdates": {
            "type": "boolean"
          },
          "preferredLanguage": {
            "type": "string"
          }
        }
      },
...
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Office Supplies"
          },
          "description": {
            "type": "string",
            "example": "Pens, paper, and staplers"
          },
          "category": {
            "enum": [
              "OFFICE",
              "TRAVEL",
              "EMPLOYEE",
              "PROPERTY",
              "UTILITIES",
              "MARKETING",
              "EQUIPMENT",
              "OTHER"
            ],
            "type": "string",
            "example": "OFFICE"
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "example": 250.5
          },
          "expenseDate": {
            "type": "string",
            "example": "2026-08-25T10:00:00Z"
          },
          "paymentMethod": {
            "enum": [
              "CASH",
              "BANK_TRANSFER",
              "CREDIT_CARD",
              "UPI",
              "CHEQUE",
              "OTHER"
            ],
            "type": "string",
            "example": "CREDIT_CARD"
          },
          "notes": {
            "type": "string",
            "example": "Bought from Staples"
          },
          "isVoided": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "title",
          "category",
          "amount",
          "expenseDate"
        ]
      },
      "UpdateExpenseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Updated Office Supplies"
          },
          "description": {
            "type": "string",
            "example": "Updated description"
          },
          "category": {
            "enum": [
              "OFFICE",
              "TRAVEL",
              "EMPLOYEE",
              "PROPERTY",
              "UTILITIES",
              "MARKETING",
              "EQUIPMENT",
              "OTHER"
            ],
            "type": "string",
            "example": "OFFICE"
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "example": 300
          },
          "expenseDate": {
            "type": "string",
            "example": "2026-08-26T10:00:00Z"
          },
          "paymentMethod": {
            "enum": [
              "CASH",
              "BANK_TRANSFER",
              "CREDIT_CARD",
              "UPI",
              "CHEQUE",
              "OTHER"
            ],
            "type": "string",
            "example": "CASH"
          },
          "notes": {
            "type": "string",
            "example": "Price went up"
          },
          "isVoided": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}